pyproject.toml
[tool.black]
line-length = 79

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyutils"
version = "1.0.0"
description = "A collection of Python utility scripts."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Miles Barr" }
]
dependencies = [
    "Pillow"
]
keywords = [
    "utilities",
    "utils",
    "scripts",
    "tools",
    "helpers"
]

classifiers = [
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]

[project.urls]
Homepage = "https://github.com/milesbarr/python-utils"

[tool.setuptools]
packages = ["pyutils"]
